split move on press
