+++ inc/structs.h
@@ -71,11 +71,11 @@ typedef struct s_olist
 	struct s_olist	*next;
 }	t_olist;
 
-typedef struct s_llist // added new 
-{
-    t_light *light;
-    struct  s_llist *next;
-} t_llist;
+// typedef struct s_llist // added new
+// {
+//     t_light *light;
+//     struct  s_llist *next;
+// } t_llist;
 
 typedef struct s_intersection
 {
@@ -115,7 +115,7 @@ typedef struct s_light
 typedef struct s_scene
 {
 	t_olist			*obj_list;
-    t_llist         *light_list; // added new 
+    // t_llist         *light_list; // added new
 	t_ambient		ambient;
 	t_light			light;
 	t_cam			cam;
@@ -125,7 +125,7 @@ typedef struct s_scene
     int             next_obj_id;
 	char			*error_line;
 	t_olist			*obj_selected;
-    t_llist         *light_selected; // added new 
+    // t_llist         *light_selected; // added new
 }					t_scene;
 
 typedef struct s_data
