+++ maps/all_obj.rt
@@ -1,6 +1,8 @@
 A 0.4 255,255,255
 C 0,2,-10 0,0,1 60
-L 10,15,-10 1.0 255,255,255
+L 10,15,-10 1.0 255,255,255 
 pl 0,-5,0 0,1,0 120,200,120
 sp ,0,20 6 255,150,150
 cy -5,0,15 0,1,0 2 6 150,150,255
+L 10,15,-10 1.0 255,255,255
+L 10,15,-10 1.0 255,255,255
