+++ maps/base.rt
@@ -3,4 +3,4 @@ C 0,1,-5 0,0,1 40
 L 0,5,-5 0.7 0,255,255
 
 pl 0,0,0 0,1,0 200,200,200
-sp 0,1,0 0 255,0,255
+sp 0,1,0 2 255,0,255
